add models Flowers, Messages and Flags
