+++ models/Messages.js
@@ -1,0 +1,40 @@
+const { Model, DataTypes } = require('sequelize');
+const sequelize = require('../config/connection');
+
+class Messages extends Model {}
+
+Messages.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true,
+      autoIncrement: true
+    },
+    message_text: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    sender_id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: 'user',
+        key: 'id'
+      }
+    },
+    recipient_id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: 'user',
+        key: 'id'
+      } 
+    }
+  },
+  {
+    sequelize,
+    modelName: 'messages',
+  });
+
+module.exports = Messages
